# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface Movie {
  id: number;
  title: string;
  overview: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  vote_count: number;
  release_date: string;
  genre_ids: number[];
  popularity: number;
}

export interface MovieListResponse {
  results: Movie[];
  page: number;
  total_pages: number;
  total_results: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface WatchlistMovieItem {
  id: number;
  tmdb_id: number;
  title: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  vote_average?: number | null;
  /** @nullable */
  release_date?: string | null;
  added_at: string;
}

export interface WatchlistResponse {
  items: WatchlistMovieItem[];
}

export interface AddToWatchlistBody {
  tmdb_id: number;
  title: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  vote_average?: number | null;
  /** @nullable */
  release_date?: string | null;
}

export interface VideoResult {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
}

export interface VideoListResponse {
  results: VideoResult[];
}

export interface GenreListResponse {
  genres: Genre[];
}

export interface MovieDetail {
  id: number;
  title: string;
  tagline: string;
  overview: string;
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  vote_average: number;
  vote_count: number;
  release_date: string;
  /** @nullable */
  runtime?: number | null;
  genres: Genre[];
  popularity: number;
  status: string;
  /** @nullable */
  homepage?: string | null;
}

export type SearchMoviesParams = {
  q: string;
  page?: number;
};

export type GetMoviesByGenreParams = {
  genre_id: number;
  page?: number;
};