# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /movies/search:
    get:
      operationId: searchMovies
      tags: [movies]
      summary: Search movies by query
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
...
  /movies/by-genre:
    get:
      operationId: getMoviesByGenre
      tags: [movies]
      summary: Get movies by genre ID
      parameters:
        - name: genre_id
          in: query
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Movies in genre
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
...
  /movies/{id}/videos:
    get:
      operationId: getMovieVideos
      tags: [movies]
      summary: Get videos (trailers) for a movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoListResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /movies/{id}:
    get:
      operationId: getMovieDetail
      tags: [movies]
      summary: Get movie detail by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Movie:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        overview:
          type: string
        poster_path:
          type: ["string", "null"]
        backdrop_path:
          type: ["string", "null"]
        vote_average:
          type: number
        vote_count:
          type: integer
        release_date:
          type: string
        genre_ids:
          type: array
          items:
            type: integer
        popularity:
          type: number
      required:
        - id
        - title
        - overview
        - vote_average
        - vote_count
        - release_date
        - genre_ids
        - popularity
    MovieListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
        page:
          type: integer
        total_pages:
          type: integer
        total_results:
          type: integer
      required:
        - results
        - page
        - total_pages
        - total_results
    Genre:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
...
    VideoResult:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        site:
          type: string
        type:
          type: string
      required:
        - id
        - key
        - name
        - site
        - type
    VideoListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/VideoResult"
      required:
        - results
...
    MovieDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        tagline:
          type: string
        overview:
          type: string
        poster_path:
          type: ["string", "null"]
        backdrop_path:
          type: ["string", "null"]
        vote_average:
          type: number
        vote_count:
          type: integer
        release_date:
          type: string
        runtime:
          type: ["integer", "null"]
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
        popularity:
          type: number
        status:
          type: string
        homepage:
          type: ["string", "null"]
      required:
        - id
        - title
        - tagline
        - overview
        - vote_average
        - vote_count
        - release_date
        - genres
        - popularity
        - status