# lib/api-zod/src/generated/types/videoResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoResult {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
}